news_summarizer/requirements.txt
requests==2.31.0
anthropic==0.45.2
langchain==0.1.16
langchain-community==0.0.34
python-dotenv==1.0.0
